pyproject.toml
[tool.poetry]
name = "tvim"
version = "0.1.0"
description = ""
authors = ["Rich Wareham <rich@richwareham.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
tvim = "tvim:main"

[tool.poetry.dependencies]
python = "^3.11"
pytvision = {git = "https://github.com/rjw57/pytvision.git"}
pynvim = "^0.5.0"
numpy = "^1.26.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
disable_error_code = ["attr-defined","name-defined"]
